[2/2] Settings: On-The-Go mode
On-The-Go turns on the camera of your phone and streams it to an
overlay,
which is drawn over everything.
this makes you feel like your display is transparent, as you see the
content
behind it.

its really useful when you are using your phone while walking, thats why
we called it "On-The-Go"

have fun!

Change-Id: If1146cdb1529655dede774b371cc55f7b782f14e
Signed-off-by: Varun <date.varun123@gmail.com>