Settings: Implement new battery options
This is the new way of accessing battery options. In battery settings,
click the 3 dot menu and you'll be presented with a new option called
"Battery options". Selecting it will take you to the battery options
screen with various options for the battery, such as custom icons,
battery text etc. The usual stuff until newer things approach.

*Java class is empty until what features are determined
*New battery strings for options
*Make Battery options searchable

Author: sirhc95@cypheros.co